[refactoring] Clean-up for chipid files
- Spelling fix in tutorial.md
- REN param description --> dev_type
- ADD param ref_manual_id
- CHNG stlink_chipid_params
- MV old chipids to chipid_db_old.h
- Updated chipid files (F1, F2, WB, WLE)
