--- config/chips/F1_Low_density_device.chip
@@ -1,13 +1,0 @@
-# Chip-ID file for F1 Low-density device
-#
-chip_id 0x412
-description F1 Low-density device
-flash_type 1
-flash_pagesize 0x400
-sram_size 0x2800
-bootrom_base 0x1ffff000
-bootrom_size 0x800
-option_base 0x1ffff800
-option_size 0x10
-flags swo
-
