--- config/chips/F1xx_High_density.chip
@@ -1,13 +1,0 @@
-# Chip-ID file for F1xx High-density
-#
-chip_id 0x414
-description F1xx High-density
-flash_type 1
-flash_pagesize 0x800
-sram_size 0x10000
-bootrom_base 0x1ffff000
-bootrom_size 0x800
-option_base 0x1ffff800
-option_size 0x10
-flags swo
-
