--- config/chips/F1xx_Medium_density.chip
@@ -1,13 +1,0 @@
-# Chip-ID file for F1xx Medium-density
-#
-chip_id 0x410
-description F1xx Medium-density
-flash_type 1
-flash_pagesize 0x400
-sram_size 0x5000
-bootrom_base 0x1ffff000
-bootrom_size 0x800
-option_base 0x1ffff800
-option_size 0x10
-flags swo
-
