+++ config/chips/F1xx_XLD.chip
@@ -1,0 +1,14 @@
+# Chip-ID file for STM32F1xx XL density device
+#
+dev_type STM32F1xx XL density device
+ref_manual_id 0008
+chip_id 0x430              // STLINK_CHIPID_STM32_F1_XLD
+flash_type 2               // STLINK_FLASH_TYPE_F1_XL
+flash_size_reg 0x1ffff7e0
+flash_pagesize 0x800
+sram_size 0x18000
+bootrom_base 0x1fffe000
+bootrom_size 0x1800
+option_base 0x0
+option_size 0x0
+flags swo
