+++ config/chips/F2xx.chip
@@ -1,13 +1,14 @@
-# Chip-ID file for F2xx
+# Chip-ID file for STM32F2xx device (STM32F205xx, STM32F207xx, STM32F215xx, STM32F217xx)
 #
-chip_id 0x411
-description F2xx
-flash_type 3
+dev_type STM32F2xx device
+ref_manual_id 0033
+chip_id 0x411                // STLINK_CHIPID_STM32_F2
+flash_type 3                 // STLINK_FLASH_TYPE_F4
+flash_size_reg 0x1fff7a22
 flash_pagesize 0x20000
 sram_size 0x20000
 bootrom_base 0x1fff0000
 bootrom_size 0x7800
-option_base 0x1fffc000
-option_size 0x4
+option_base 0x1fffc000       // STM32_F2_OPTION_BYTES_BASE
+option_size 0x4              // 4
 flags swo
-
