--- config/chips/WB5x_3x.chip
@@ -1,13 +1,0 @@
-# Chip-ID file for WB5x/3x
-#
-chip_id 0x495
-description WB5x/3x
-flash_type 9
-flash_pagesize 0x1000
-sram_size 0x40000
-bootrom_base 0x1fff0000
-bootrom_size 0x7000
-option_base 0x0
-option_size 0x0
-flags swo
-
