+++ config/chips/WLEx.chip
@@ -1,13 +1,14 @@
-# Chip-ID file for WLEx
+# Chip-ID file for STM32WLEx device
 #
-chip_id 0x497
-description WLEx
-flash_type 9
-flash_pagesize 0x800
+dev_type STM32WLEx device
+ref_manual_id 0033
+chip_id 0x497                // STLINK_CHIPID_STM32_WLE
+flash_type 9                 // STLINK_FLASH_TYPE_WB
+flash_size_reg 0x1fff75e0
+flash_pagesize 0x800         // 2k
 sram_size 0x10000
 bootrom_base 0x1fff0000
 bootrom_size 0x7000
 option_base 0x0
 option_size 0x0
 flags swo
-
