+++ src/st-info/info.c
@@ -42,7 +42,7 @@ static void stlink_print_info(stlink_t *sl) {
     printf("  chipid:     0x%.4x\n", sl->chip_id);
 
     params = stlink_chipid_get_params(sl->chip_id);
-    if (params) { printf("  descr:      %s\n", params->description); }
+    if (params) { printf("  dev-type:   %s\n", params->dev_type); }
 }
 
 static void stlink_probe(enum connect_type connect, int freq) {
@@ -118,7 +118,7 @@ static int print_data(int ac, char **av) {
         const struct stlink_chipid_params *params = stlink_chipid_get_params(sl->chip_id);
         if (params == NULL) { return(-1); }
 
-        printf("%s\n", params->description);
+        printf("%s\n", params->dev_type);
     }
 
     if (sl) {
